fix : 파일 이름 오타 수정 및 CommonProductList의 title 넓이 버그 수정
