+++ src/pages/used-market/Index.jsx
@@ -1,13 +1,13 @@
 import styles from './Index.module.css';
 import BestProductList from './model/BestProductList';
-import CommonPorductList from './model/CommonPorductList';
+import CommonProductList from './model/CommonProductList';
 
 const UsedMarket = () => {
   return (
     <section>
       <div className={styles.used_market__container}>
         <BestProductList />
-        <CommonPorductList />
+        <CommonProductList />
       </div>
     </section>
   );
