+++ src/pages/used-market/model/CommonProductList.module.css
@@ -1,0 +1,163 @@
+.header {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.mobile_title_wrapper {
+  display: flex;
+  justify-content: space-between;
+
+  width: 100%;
+
+  margin-bottom: 10px;
+}
+
+.title {
+  font-size: 1.4rem;
+  font-weight: bold;
+
+  align-self: center;
+}
+
+.actions {
+  display: flex;
+  justify-content: flex-end;
+
+  width: 100%;
+  height: 42px;
+
+  gap: 10px;
+}
+
+.search_input {
+  width: 100%;
+  height: 100%;
+
+  background-color: #f3f4f6;
+
+  border: none;
+  border-radius: 12px;
+  padding: 20px;
+}
+
+.search_input:focus {
+  outline: none;
+}
+
+.register_button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 150px;
+  height: 100%;
+
+  color: white;
+  background-color: var(--primary-color);
+
+  border: none;
+  border-radius: 8px;
+  padding: 15px;
+
+  cursor: pointer;
+}
+
+.order_dropdown {
+  position: relative;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 60px;
+  height: 100%;
+
+  background-color: white;
+
+  border: 1px solid #e5e7eb;
+  border-radius: 8px;
+  padding: 20px;
+
+  cursor: pointer;
+}
+
+.order_dropdown span {
+  font-size: 1rem;
+}
+
+.order_dropdown svg {
+  font-size: 1.2rem;
+}
+
+.order_dropdown_list {
+  position: absolute;
+  background-color: white;
+
+  right: 20px;
+
+  border-radius: 12px;
+  border: 1px solid #e5e7eb;
+
+  margin-top: 10px;
+
+  z-index: 1;
+}
+
+.order_dropdown_list li {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  height: 42px;
+  width: 120px;
+
+  cursor: pointer;
+}
+
+.list {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 20px;
+
+  margin-top: 20px;
+}
+
+@media screen and (min-width: 820px) {
+  .header {
+    flex-direction: row;
+  }
+
+  .title {
+    font-size: 1.4rem;
+  }
+
+  .actions {
+    width: 50%;
+  }
+
+  .search_input {
+    width: 330px;
+  }
+
+  .order_dropdown {
+    width: 120px;
+    justify-content: space-between;
+  }
+
+  .order_dropdown_list {
+    position: relative;
+    right: 0;
+  }
+
+  .list {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+
+@media screen and (min-width: 1200px) {
+  .list {
+    grid-template-columns: repeat(5, 1fr);
+  }
+}
